requirements.txt
python-frontmatter
Pillow

